Auth Controller methods

+++ app/Http/Controllers/PostController.php
@@ -5,9 +5,11 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Requests\PostFormRequest;
 use App\Models\Post;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class PostController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Show the form for creating a new resource.
      */
@@ -43,9 +45,11 @@ class PostController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $id)
+    public function edit(Post $post)
     {
-        return view('posts.edit', ['post' => Post::findOrFail($id)]);
+        $this->authorize('update', $post);
+
+        return view('posts.edit', ['post' => $post]);
     }
 
     /**
@@ -53,6 +57,7 @@ class PostController extends Controller
      */
     public function update(PostFormRequest $request, Post $post)
     {
+        $this->authorize('update', $post);
         $validated = $request->validated();
 
         $post->update($validated);
@@ -67,10 +72,9 @@ class PostController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Post $post)
     {
-        $post = Post::findOrFail($id);
-
+        $this->authorize('update', $post);
         $post->delete();
 
         return redirect()
